Move applyDefaults to external util & adjust tests
