+++ lib/processors/bundlers/moduleBundler.js
@@ -153,8 +153,7 @@ export default function({resources, taskUtil, options: {
 		ignoreMissingModules: false
 	}, bundleOptions);
 
-	// Apply defaults without modifying the passed object
-	bundleDefinition = applyDefaultsToBundleDefinition(bundleDefinition, taskUtil);
+	// bundleDefinition's defaults get applied in the corresponding standard tasks
 
 	const pool = new LocatorResourcePool({
 		ignoreMissingModules: bundleOptions.ignoreMissingModules
@@ -195,27 +194,4 @@ export default function({resources, taskUtil, options: {
 				}
 			}));
 		});
-}
-
-function applyDefaultsToBundleDefinition(bundleDefinition, taskUtil) {
-	bundleDefinition.sections = bundleDefinition?.sections?.map((section) => {
-		const defaultValues = {
-			resolve: false,
-			resolveConditional: false,
-			renderer: false,
-			sort: true,
-			declareRawModules: false,
-		};
-
-		// Since specVersion: 4.0 "require" section starts loading asynchronously.
-		// If specVersion cannot be determined, the latest spec is taken into account.
-		// This is a breaking change in specVersion: 4.0
-		if (section.mode === "require" && (!taskUtil || taskUtil.getProject().getSpecVersion().gte("4.0"))) {
-			defaultValues.async = true;
-		}
-
-		return Object.assign(defaultValues, section);
-	});
-
-	return bundleDefinition;
 }
