+++ lib/tasks/bundlers/generateLibraryPreload.js
@@ -1,6 +1,7 @@
 import {getLogger} from "@ui5/logger";
 const log = getLogger("builder:tasks:bundlers:generateLibraryPreload");
 import moduleBundler from "../../processors/bundlers/moduleBundler.js";
+import {applyDefaultsToBundleDefinition} from "./utils/applyDefaultsToBundleDefinition.js";
 import {negateFilters} from "../../lbt/resources/ResourceFilterList.js";
 import createModuleNameMapping from "./utils/createModuleNameMapping.js";
 
@@ -265,7 +266,8 @@ export default async function({workspace, taskUtil, options: {skipBundles = [], 
 		if (coreVersion) {
 			options.targetUi5CoreVersion = coreVersion;
 		}
-		return moduleBundler({options, resources, taskUtil});
+		options.bundleDefinition = applyDefaultsToBundleDefinition(options.bundleDefinition, taskUtil);
+		return moduleBundler({options, resources});
 	};
 
 	return nonDbgWorkspace.byGlob("/**/*.{js,json,xml,html,properties,library,js.map}").then(async (resources) => {
